chore: relocate operator mock file
